UEPyEngine, UEPyEditor, UEPyModule: Added console exec function to UEPyEngine and removed it from UEPyEditor so that it's usable in standalone scripts.
Fix various build ordering/ifdef issues for shipping
Add clear_event function for multicasting
