+++ Source/UnrealEnginePython/Private/UEPyEditor.h
@@ -23,8 +23,6 @@ PyObject *py_unreal_engine_set_fbx_import_option(PyObject *, PyObject *);
 
 PyObject *py_unreal_engine_create_modal_save_asset_dialog(PyObject *, PyObject *);
 
-PyObject *py_unreal_engine_console_exec(PyObject *, PyObject * args);
-
 PyObject *py_unreal_engine_editor_tick(PyObject *, PyObject *);
 
 PyObject *py_unreal_engine_get_discovered_plugins(PyObject *, PyObject *);
